Misc thread fixes

+++ src/modules/playground.ts
@@ -20,8 +20,13 @@ import {
 	truncate,
 } from '../util/codeBlocks';
 import { LimitedSizeMap } from '../util/limitedSizeMap';
-import { addMessageOwnership, sendWithMessageOwnership } from '../util/send';
+import {
+	addMessageOwnership,
+	getResponseChannel,
+	sendWithMessageOwnership,
+} from '../util/send';
 import fetch from 'node-fetch';
+import { isHelpChannel } from './helpthread';
 
 const LINK_SHORTENER_ENDPOINT = 'https://tsplay.dev/api/short';
 const MAX_EMBED_LENGTH = 512;
@@ -64,13 +69,14 @@ export class PlaygroundModule extends Module {
 		const exec = PLAYGROUND_REGEX.exec(msg.content);
 		if (!exec) return;
 		const embed = createPlaygroundEmbed(msg.author, exec);
-		if (exec[0] === msg.content) {
+		if (exec[0] === msg.content && !isHelpChannel(msg.channel)) {
 			// Message only contained the link
 			await sendWithMessageOwnership(msg, { embeds: [embed] });
 			await msg.delete();
 		} else {
 			// Message also contained other characters
-			const botMsg = await msg.channel.send({
+			const channel = await getResponseChannel(msg);
+			const botMsg = await channel.send({
 				embeds: [embed],
 				content: `${msg.author} Here's a shortened URL of your playground link! You can remove the full link from your message.`,
 			});
